requirements.txt
Flask
Flask-Session
PyYAML